# lib/api-zod/src/generated/types/citation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Citation {
  id: string;
  documentId: string;
  filename: string;
  page: number;
  chunk: number;
  score: number;
  excerpt: string;
}